+++ MovBlokGame/UnifyEngine/Vector2.java
@@ -1,14 +1,14 @@
 package UnifyEngine;
 
 public class Vector2 {
-	public int x,y;
+	public float x,y;
 	
 	public Vector2() {
-		this.x=0;
-		this.y=0;
+		this.x = 0;
+		this.y = 0;
 	}
 	
-	public Vector2(int _x, int _y) {
+	public Vector2(float _x, float _y) {
 		this.x = _x;
 		this.y = _y;
 	}
